# web/openapi.json
{
  "components": {
    "schemas": {
...
      "AttemptResponse": {
        "description": "One upstream attempt. SPEC \u00a710.2's ``failover_attempts`` entry, as JSON.",
        "properties": {
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "model_name": {
            "title": "Model Name",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "target_id": {
            "format": "uuid",
            "title": "Target Id",
            "type": "string"
          }
        },
        "required": [
          "target_id",
          "model_name",
          "status",
          "latency_ms",
          "retryable"
        ],
        "title": "AttemptResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_RequestLogResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RequestLogResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[RequestLogResponse]",
        "type": "object"
      },
...
      "RequestDetailResponse": {
        "additionalProperties": false,
        "description": "One request, with everything the \u00a710.3 drawer draws.\n\nThe retrieval lists are here and empty until task 10 fills them, for the same reason\nthe gateway editor shows its unbuilt sections: a drawer that grows a panel later moves\neverything the reader has learned the position of.\n\n``failover_attempts`` is empty for the overwhelming majority of requests, and that is\ninformation rather than an omission: it means one target answered, which the ``log``\nfields already describe completely. Non-empty means more than one was involved.",
        "properties": {
          "failover_attempts": {
            "items": {
              "$ref": "#/components/schemas/AttemptResponse"
            },
            "title": "Failover Attempts",
            "type": "array"
          },
          "log": {
            "$ref": "#/components/schemas/RequestLogResponse"
          },
          "retrieved_chunk_ids": {
            "items": {},
            "title": "Retrieved Chunk Ids",
            "type": "array"
          },
          "retrieved_fact_ids": {
            "items": {},
            "title": "Retrieved Fact Ids",
            "type": "array"
          },
          "transcript": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TranscriptResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "log",
          "retrieved_chunk_ids",
          "retrieved_fact_ids",
          "failover_attempts"
        ],
        "title": "RequestDetailResponse",
        "type": "object"
      },
      "RequestLogResponse": {
        "additionalProperties": false,
        "description": "One row of the request table. No bodies \u2014 that is what the detail view is for.",
        "properties": {
          "api_key_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key Id"
          },
          "bodies_omitted": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bodies Omitted"
          },
          "completion_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Tokens"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "end_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End User Id"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "failed_after_stream_start": {
            "title": "Failed After Stream Start",
            "type": "boolean"
          },
          "gateway_id": {
            "format": "uuid",
            "title": "Gateway Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latency_retrieval_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Retrieval Ms"
          },
          "latency_total_ms": {
            "title": "Latency Total Ms",
            "type": "integer"
          },
          "latency_ttft_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Ttft Ms"
          },
          "latency_upstream_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Upstream Ms"
          },
          "memory_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Tokens"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "prompt_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Tokens"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "response_truncated": {
            "title": "Response Truncated",
            "type": "boolean"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "status_code": {
            "title": "Status Code",
            "type": "integer"
          },
          "streamed": {
            "title": "Streamed",
            "type": "boolean"
          },
          "upstream_model_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upstream Model Id"
          }
        },
        "required": [
          "id",
          "created_at",
          "gateway_id",
          "api_key_id",
          "end_user_id",
          "session_id",
          "upstream_model_id",
          "model_name",
          "status_code",
          "error_code",
          "error_message",
          "streamed",
          "latency_total_ms",
          "latency_retrieval_ms",
          "latency_ttft_ms",
          "latency_upstream_ms",
          "prompt_tokens",
          "completion_tokens",
          "memory_tokens",
          "request_id",
          "response_truncated",
          "failed_after_stream_start",
          "bodies_omitted"
        ],
        "title": "RequestLogResponse",
        "type": "object"
      },
...
      "TranscriptResponse": {
        "additionalProperties": false,
        "description": "The bodies, as far as they were stored.",
        "properties": {
          "assembled_prompt": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assembled Prompt"
          },
          "distilled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distilled At"
          },
          "request_body": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Body"
          },
          "response_body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Body"
          }
        },
        "title": "TranscriptResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Memory Gateway",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/logs": {
      "get": {
        "description": "The request table: metadata only, newest first, cursor-paginated.\n\nCursor rather than offset because this list is being written to while it is read \u2014\nSPEC \u00a712.2's reason, and the one screen where an offset page would visibly duplicate\nand skip rows as traffic arrives.",
        "operationId": "list_logs_api_v1_logs_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          },
          {
            "in": "query",
            "name": "gateway_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Gateway Id"
            }
          },
          {
            "in": "query",
            "name": "upstream_model_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Upstream Model Id"
            }
          },
          {
            "in": "query",
            "name": "status_class",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 3,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Class"
            }
          },
          {
            "in": "query",
            "name": "end_user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End User Id"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "in": "query",
            "name": "streamed",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Streamed"
            }
          },
          {
            "in": "query",
            "name": "min_latency_ms",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "minimum": 0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Latency Ms"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 200,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RequestLogResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Logs",
        "tags": [
          "monitoring"
        ]
      }
    },
    "/api/v1/logs/{log_id}": {
      "get": {
        "description": "One request in full, including whatever of the transcript was stored.\n\nNo time range: the id is a UUIDv7 and carries the millisecond it was minted, so the\nserver works out which day's partition to look in. A 404 covers \"no such request\" and\n\"another organization's request\" alike.",
        "operationId": "get_log_api_v1_logs__log_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "log_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Log Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Log",
        "tags": [
          "monitoring"
        ]
      }
    },
...
  }
}